add cross_attention
